+++ noxfile.py
@@ -17,7 +17,8 @@ def pre_commit(session: nox.Session) -> None:
     session.run("pre-commit", "run", "--all-files", "--show-diff-on-failure")
 
 
-@nox.session(python=["pypy3.7", "pypy3.8", "3.7", "3.8", "3.9", "3.10", "3.11"])
+# Keep list in-sync with ci.yml/test-linux & pyproject.toml
+@nox.session(python=["pypy3.8", "3.8", "3.9", "3.10", "3.11", "3.12"])
 def tests(session: nox.Session) -> None:
     session.install("Twisted", "coverage[toml]")
     posargs = list(session.posargs)
@@ -37,8 +38,6 @@ def tests(session: nox.Session) -> None:
 
     if os.environ.get("CI") != "true":
         session.notify("coverage_report")
-    else:
-        session.run("coverage", "combine")
 
 
 @nox.session
@@ -56,8 +55,16 @@ def check_newsfragment(session: nox.Session) -> None:
 
 
 @nox.session
+def draft_newsfragment(session: nox.Session) -> None:
+    session.install(".")
+    session.run("python", "-m", "towncrier.build", "--draft")
+
+
+@nox.session
 def typecheck(session: nox.Session) -> None:
-    session.install(".", "mypy", "types-setuptools")
+    # Click 8.1.4 is bad type hints -- lets not complicate packaging and only
+    # pin here.
+    session.install(".", "mypy", "click!=8.1.4")
     session.run("mypy", "src")