+++ src/towncrier/_shell.py
@@ -11,11 +11,10 @@ from __future__ import annotations
 
 import click
 
-from click_default_group import DefaultGroup
-
 from ._version import __version__
 from .build import _main as _build_cmd
 from .check import _main as _check_cmd
+from .click_default_group import DefaultGroup
 from .create import _main as _create_cmd